Drivers: refactor, add haloCheckWebNFCPermission
