+++ api/web.js
@@ -4,12 +4,11 @@
  * License: MIT
  */
 
-const {
-    execHaloCmdWeb, detectMethod
-} = require("../drivers/common");
 const {HaloGateway} = require("../halo/gateway/requestor");
 const {haloFindBridge} = require("../web/web_utils");
 const {haloGateExecutorCreateWs, haloGateExecutorUserConfirm} = require("../halo/gateway/executor");
+const {execHaloCmdWeb, detectMethod} = require("../drivers/web");
+const {checkWebNFCPermission} = require("../drivers/webnfc");
 
 /**
  * The LibHaLo stable API. Please don't depend on the functions imported from anywhere else
@@ -20,6 +19,7 @@ module.exports = {
     execHaloCmdWeb,
     haloFindBridge,
     haloGetDefaultMethod: detectMethod,
+    haloCheckWebNFCPermission: checkWebNFCPermission,
 
     // for web usage with gateway
     HaloGateway,
