+++ drivers/common.js
@@ -4,11 +4,7 @@
  * License: MIT
  */
 
-const {execCredential} = require("./credential");
-const {execWebNFC} = require("./webnfc");
 const {
-    NFCAbortedError,
-    NFCMethodNotSupported,
     HaloLogicError,
     HaloTagError
 } = require("../halo/exceptions");
@@ -17,23 +13,6 @@ const {
     cmdSignChallenge
 } = require("../halo/commands");
 const {ERROR_CODES} = require("../halo/errors");
-
-let isCallRunning = null;
-
-/**
- * Detect the best command execution method for the current device.
- * @returns {string} Either "credential" or "webnfc".
- */
-function detectMethod() {
-    try {
-        new NDEFReader();
-    } catch (e) {
-        // WebNFC not supported
-        return "credential";
-    }
-
-    return "webnfc";
-}
 
 async function execHaloCmd(command, options) {
     command = Object.assign({}, command);
@@ -65,18 +44,6 @@ async function execHaloCmd(command, options) {
     }
 }
 
-function makeDefault(curValue, defaultValue) {
-    if (typeof curValue === "undefined") {
-        return defaultValue;
-    }
-
-    if (curValue === null) {
-        return defaultValue;
-    }
-
-    return curValue;
-}
-
 function checkErrors(res) {
     if (res.length === 2 && res[0] === 0xE1) {
         if (ERROR_CODES.hasOwnProperty(res[1])) {
@@ -88,59 +55,7 @@ function checkErrors(res) {
     }
 }
 
-/**
- * Execute the NFC command from the web browser.
- * @param command Command specification object.
- * @param options Additional options for the command executor.
- * @returns {Promise<*>} Command execution result.
- */
-async function execHaloCmdWeb(command, options) {
-    if (options && !options.noDebounce && isCallRunning) {
-        throw new NFCAbortedError("The operation was debounced.");
-    }
-
-    isCallRunning = true;
-
-    options = options ? Object.assign({}, options) : {};
-    options.method = makeDefault(options.method, detectMethod());
-    options.noDebounce = makeDefault(options.noDebounce, false);
-    options.compatibleCallMode = makeDefault(options.compatibleCallMode, true);
-
-    command = command ? Object.assign({}, command) : {};
-
-    try {
-        let cmdOpts = {};
-
-        if (options.method === "credential") {
-            cmdOpts = {
-                method: "credential",
-                exec: async (command) => await execCredential(command, {
-                    debugCallback: options.debugCallback,
-                    statusCallback: options.statusCallback,
-                    compatibleCallMode: options.compatibleCallMode
-                })
-            };
-        } else if (options.method === "webnfc") {
-            cmdOpts = {
-                method: "webnfc",
-                exec: async (command) => await execWebNFC(command, {
-                    debugCallback: options.debugCallback,
-                    statusCallback: options.statusCallback
-                })
-            };
-        } else {
-            throw new NFCMethodNotSupported("Unsupported options.method parameter specified.");
-        }
-
-        return await execHaloCmd(command, cmdOpts);
-    } finally {
-        isCallRunning = false;
-    }
-}
-
 module.exports = {
-    execHaloCmdWeb,
     execHaloCmd,
-    checkErrors,
-    detectMethod
+    checkErrors
 };
